requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
astroid==2.4.2
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.3
channels==1.1.8
colorama==0.4.4
constantly==15.1.0
cryptography==3.2.1
daphne==1.4.2
Django==2.2
django-webpack-loader==0.7.0
djangorestframework==3.12.2
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack-python==0.5.6
psycopg2==2.8.6
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pytz==2020.4
redis==2.10.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
wrapt==1.12.1
zope.interface==5.2.0
